etl/requirements.txt
python-dotenv==0.20.0
psycopg2-binary==2.9.1
elasticsearch==7.7.0
redis==4.3.1
requests==2.27.1
Faker==13.13.0
